created laser prefab, however there are issues while trying to fire the laser

+++ Assets/Scripts/Position.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class Position : MonoBehaviour {
 	//public Vector3 vector3;
-	public int radius = 1;
+	public float radius = 1;
 
 	//THIS METHOD DRAWS A SPHERE IN A POSITION
 	void OnDrawGizmos(){
